mcp-servers/supabase-schema-server/requirements.txt
psycopg2-binary==2.9.10
mcp==1.0.0
python-dotenv==1.0.0
